+++ core/src/files.js
@@ -9,6 +9,7 @@ export * from './projects.js';
 export * from './guidance.js';
 export * from './memory.js';
 export * from './artifacts.js';
+export * from './packets.js';
 export * from './schema.js';
 export { ulid } from './ids.js';
 export { slugify } from './slug.js';